service_a/requirements.txt
# FastAPI and Uvicorn for running the application
fastapi==0.114.0
uvicorn==0.30.6

# OpenTelemetry core and instrumentation for FastAPI and Requests
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-requests==0.48b0
## The following requirements were added by pip freeze:
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
exceptiongroup==1.2.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
h11==0.14.0
idna==3.8
importlib_metadata==8.4.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
protobuf==4.25.4
pydantic==2.9.1
pydantic_core==2.23.3
requests==2.32.3
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
zipp==3.20.1